delayed init of lion code
