+++ pyterrier_dr/__init__.py
@@ -11,7 +11,7 @@ from pyterrier_dr.electra import ElectraScorer
 from pyterrier_dr.star import STAR
 from pyterrier_dr.repllama import RepLLama
 from pyterrier_dr.dragon import Dragon
-#from pyterrier_dr.lion import LionLlamaDense
+from pyterrier_dr.lion import LionLlamaDense
 from pyterrier_dr.bge_m3 import BGEM3, BGEM3QueryEncoder, BGEM3DocEncoder
 from pyterrier_dr.cde import CDE, CDECache
 from pyterrier_dr.prf import AveragePrf, VectorPrf
@@ -21,7 +21,8 @@ from pyterrier_dr._mmr import MmrScorer
 __all__ = ["FlexIndex", "DocnoFile", "NilIndex", "NumpyIndex", "RankedLists", "FaissFlat", "FaissHnsw", "MemIndex", "TorchIndex",
            "BiEncoder", "BiQueryEncoder", "BiDocEncoder", "BiScorer", "HgfBiEncoder", "TasB", "RetroMAE", "SBertBiEncoder", "Ance",
            "Query2Query", "GTR", "E5", "TctColBert", "ElectraScorer", "LightningIRMonoScorer", "BGEM3", "BGEM3QueryEncoder", "BGEM3DocEncoder", "CDE", "CDECache",
-           "STAR", #"LionLlamaDense",
+           "STAR", 
+           "LionLlamaDense",
            "Dragon",
            "RepLLama",
            "SimFn", "infer_device", "AveragePrf", "VectorPrf", "ILS", "ils", "MmrScorer"]
