+++ pyterrier_dr/lion.sh
@@ -1,0 +1,3 @@
+/opt/miniconda3/envs/py312_lion/bin/python /root/nfs/jpq/pyterrier_dr_jpq/pyterrier_dr/jpq/run.py --pq-impl faiss2opq --pq-sample-size 50000 --model-name lion --lambda-rank --jpq-negs 200 --frozen-query-encoder --base-index /root/nfs/jpq/indices/lion --target-dir /root/nfs/jpq/experiments/repllama --M 128 --nbits 8 --in-batch-negs  2>&1 | tee /root/nfs/jpq/experiments//lion/m128_nbits8_ibn_pqss50k-faiss2opq-lr-200jpqegs.log
+/opt/miniconda3/envs/py312_lion/bin/python /root/nfs/jpq/pyterrier_dr_jpq/pyterrier_dr/jpq/run.py --pq-impl faiss2opq --pq-sample-size 50000 --model-name lion --lambda-rank --jpq-negs 200 --frozen-query-encoder --base-index /root/nfs/jpq/indices/lion --target-dir /root/nfs/jpq/experiments/repllama --M 256 --nbits 8 --in-batch-negs  2>&1 | tee /root/nfs/jpq/experiments//lion/m256_nbits8_ibn_pqss50k-faiss2opq-lr-200jpqegs.log
+/opt/miniconda3/envs/py312_lion/bin/python /root/nfs/jpq/pyterrier_dr_jpq/pyterrier_dr/jpq/run.py --pq-impl faiss2opq --pq-sample-size 50000 --model-name lion --lambda-rank --jpq-negs 200 --frozen-query-encoder --base-index /root/nfs/jpq/indices/lion --target-dir /root/nfs/jpq/experiments/repllama --M 512 --nbits 8 --in-batch-negs  2>&1 | tee /root/nfs/jpq/experiments//lion/m512_nbits8_ibn_pqss50k-faiss2opq-lr-200jpqegs.log
